+++ src/withings_cache.clj
@@ -1,10 +1,11 @@
-#!/usr/bin/env bb
 (ns src.withings-cache
   (:require
    [babashka.curl :as curl]
    [babashka.pods :as pods]
    [cheshire.core :as json]
-   [clojure.math  :refer [pow]]))
+   [clojure.java.shell :refer [sh]]
+   [clojure.math  :refer [pow]]
+   [clojure.string :as str]))
 
 (pods/load-pod 'org.babashka/mysql "0.1.1")
 (require '[pod.babashka.mysql :as mysql])
@@ -15,15 +16,13 @@
          :user     (System/getenv "MYSQL_USER")
          :password (System/getenv "MYSQL_PASSWORD")})
 
-;; debug
-;;(def wc "http://localhost:3000")
 (def wc "https://wc.kohhoh.jp")
 (def cookie "cookie.txt")
 
-;; withing-client
 (def admin    (System/getenv "WC_LOGIN"))
 (def password (System/getenv "WC_PASSWORD"))
 
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; authentication
 (defn login
   "login. if success, updates cookie and returns 302."
@@ -33,20 +32,26 @@
     (curl/post api {:raw-args ["-c" cookie "-d" params]
                     :follow-redirects false})))
 
-(comment
-  (:status (login))
-  :rcf)
+(defn login-success?
+  []
+  (= 302 (:status (login))))
+
+;; `fetch-users` requires login.
+(login-success?)
 
 ;; short-hand functions
+;; conversations to wc.kohhoh.jp require cookie.
 (defn curl-get [url & params]
   (curl/get url {:raw-args (vec (concat ["-b" cookie] params))}))
 
 (defn curl-post [url & params]
   (curl/post url {:raw-args (vec (concat ["-b" cookie] params))}))
 
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; users
 (defn fetch-users
-  "fetch users via withing-client, return the user json"
+  "fetch users via withing-client,
+   return the users data in json format."
   []
   (-> (curl-get (str wc "/api/users"))
       :body
@@ -54,37 +59,87 @@
       vec))
 
 (comment
-  (def users (fetch-users))
-  users
-  )
+  (fetch-users)
+  :rcf)
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; tokens
 (defn refresh!
   "Refresh user id's refresh token."
   [id]
   (curl-post (str wc "/api/token/" id "/refresh")))
 
-(comment
-  (refresh! 16)
-  (refresh! 51)
-  :rcf)
-
-;; (defn refresh-all!
-;;   "Refresh all available user tokens."
-;;   []
-;;   (curl-post (str wc "/api/tokens/refresh-all")))
-
 ;; pmap でスピードアップ。
-(defn refresh-all-pmap!
-  [users]
-  (->> (filter :valid users)
-       (map :id)
-       (pmap refresh!)))
+(defn refresh-all!
+  "use old users map internaly, returns refreshed user map.
+   becore fetch-users, login is required."
+  []
+  (and
+   (login-success?)
+   (->> (filter :valid (fetch-users))
+        (map :id)
+        (pmap refresh!))
+   ;; FIXME: this returns old users.
+   #_(fetch-users)))
 
 (comment
-  (refresh-all-pmap! (fetch-users))
+  (def users (refresh-all!))
   :rcf)
 
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; get meas
+;; direct download from withings
+;; need users
+(def withings "https://wbsapi.withings.net/measure")
+
+(defn to-unix-time
+  [datetime]
+  (-> (sh "date" "-d" datetime "+%s")
+      :out
+      str/trim-newline))
+
+(defn from-unix-time
+  [n]
+  (-> (sh "date" "-d" (str "@" n) "+%Y/%m/%d %T")
+      :out
+      str/trim-newline))
+
+(defn access-token [id users]
+  (->> users
+       (filter #(= id (:id %)))
+       first
+       :access))
+
+(defn withings-get-meas
+  "users must be set before calling this function."
+  [id date users]
+  (let [token  (access-token id users)
+        params (str "action=getmeas&lastupdate=" (to-unix-time date))]
+    (println "id:" id)
+    (println "access token:" token)
+    (println "params:" params)
+    (->
+     (curl/get withings
+               {:raw-args ["-H" (str "Authorization: Bearer " token)
+                           "-d" params]})
+     :body
+     (json/parse-string true)
+     (get-in [:body :measuregrps]))))
+
+(defn withings-test
+  []
+  (let [_ (refresh-all!)
+        users (fetch-users)]
+    [(withings-get-meas 51 "2022-09-01 00:00:00" users)
+     (withings-get-meas 51 "2022-12-20 00:00:00" users)]))
+
+(comment
+  (def ret (withings-test))
+  :rcf)
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; get meas
+;; get via https://wc.kohhoh.jp
 (defn get-meas-with
   [params]
   (println "get-meas-with" params)
@@ -109,12 +164,6 @@
   ([id startdate enddate]
    (get-meas id 1 startdate enddate)))
 
-(comment
-  (login)
-  (get-meas 27 "1,3,5" "2022-12-01") ;; 不適切な番号でもフェッチする。
-  (get-weight 16 "2022-10-31" "2022-11-20")
-  :rcf)
-
 ;; WITHINGS が間違い！
 ;; meastypes ではなく、meastype
 ;; meastypes=1,4,12 は invalid parameter error を起こす。
@@ -130,22 +179,19 @@
 
 (comment
   (login)
+  (def users (refresh-all!))
   (get-weight 27 "2022-12-01")
   (def record (get-meas-all 27 "2022-12-01")) ;; fixed at 0.15.5
-  record
-  (count record)
   :rcf)
 
-;;;;;;;;;;;;;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; save meas
 ;; FIXME: meas->float returns 93.60000000000001
 (defn meas->float
-  [{:keys [value unit] :as params}]
-  ;; (println value unit params)
+  [{:keys [value unit]}]
   (* value (pow 10 unit)))
 
-;; FIXME: 複数の meas には対応していない。
-;; FIXME: must uniq agains (id, created)
+;; FIXME: must uniq against (id, created)
 (defn save-one!
   [id {:keys [created measures]}]
   (let [meas (first measures) ;; <-
@@ -185,8 +231,7 @@
 (defn init-db-weight
   "init by weights data."
   [last-update]
-  (let [users (filter :valid (fetch-users))]
-    (refresh-all-pmap! users)
+  (let [users (refresh-all!)]
     (doseq [{:keys [id]} users]
       (println id)
       ;; should change to `save-multi!`
@@ -195,9 +240,3 @@
 (comment
   (init-db-weight "2022-09-01")
   :rcf)
-
-;; 48
-;; 32
-;; 17
-;; ; clojure.lang.ExceptionInfo: babashka.curl: status 400 src.withings-cache /Volumes/RAM_DISK/withings-cache/src/withings_cache.clj:44:3
-;; ; Evaluation of file withings_cache.clj failed: class clojure.lang.ExceptionInfo